+++ Runtime/TomlSerializer.Deserialize.cs
@@ -1,0 +1,7 @@
+namespace UnderLogic.Serialization.Toml
+{
+    public static partial class TomlSerializer
+    {
+        
+    }
+}